Split multi class file into multiple files and put them in a new subpackage
